AI - refactored packages
