+++ src/games/strategy/triplea/ai/proAI/data/ProPurchaseOption.java
@@ -1,4 +1,4 @@
-package games.strategy.triplea.ai.proAI;
+package games.strategy.triplea.ai.proAI.data;
 
 import java.util.ArrayList;
 import java.util.HashMap;
