+++ src/games/strategy/triplea/ai/proAI/util/ProBattleUtils.java
@@ -8,10 +8,10 @@ import games.strategy.engine.data.Unit;
 import games.strategy.engine.data.UnitType;
 import games.strategy.triplea.Properties;
 import games.strategy.triplea.ai.proAI.ProAI;
-import games.strategy.triplea.ai.proAI.ProTerritory;
-import games.strategy.triplea.ai.proAI.ProBattleResult;
-import games.strategy.triplea.ai.proAI.ProPlaceTerritory;
-import games.strategy.triplea.ai.proAI.ProPurchaseTerritory;
+import games.strategy.triplea.ai.proAI.data.ProBattleResult;
+import games.strategy.triplea.ai.proAI.data.ProPlaceTerritory;
+import games.strategy.triplea.ai.proAI.data.ProPurchaseTerritory;
+import games.strategy.triplea.ai.proAI.data.ProTerritory;
 import games.strategy.triplea.ai.proAI.logging.ProLogger;
 import games.strategy.triplea.delegate.BattleCalculator;
 import games.strategy.triplea.delegate.DiceRoll;
