+++ src/games/strategy/triplea/ai/proAI/util/ProMoveOptionsUtils.java
@@ -9,10 +9,10 @@ import games.strategy.engine.data.UnitType;
 import games.strategy.triplea.Properties;
 import games.strategy.triplea.TripleAUnit;
 import games.strategy.triplea.ai.proAI.ProAI;
-import games.strategy.triplea.ai.proAI.ProTransport;
-import games.strategy.triplea.ai.proAI.ProTerritory;
-import games.strategy.triplea.ai.proAI.ProBattleResult;
-import games.strategy.triplea.ai.proAI.ProMoveOptions;
+import games.strategy.triplea.ai.proAI.data.ProBattleResult;
+import games.strategy.triplea.ai.proAI.data.ProMoveOptions;
+import games.strategy.triplea.ai.proAI.data.ProTerritory;
+import games.strategy.triplea.ai.proAI.data.ProTransport;
 import games.strategy.triplea.ai.proAI.logging.ProLogger;
 import games.strategy.triplea.attatchments.TerritoryAttachment;
 import games.strategy.triplea.attatchments.UnitAttachment;
