+++ src/games/strategy/triplea/ai/proAI/util/ProTransportUtils.java
@@ -14,8 +14,8 @@ import games.strategy.engine.data.Territory;
 import games.strategy.engine.data.Unit;
 import games.strategy.triplea.ai.AdvancedUtils;
 import games.strategy.triplea.ai.proAI.ProAI;
-import games.strategy.triplea.ai.proAI.ProTerritory;
-import games.strategy.triplea.ai.proAI.ProPurchaseOption;
+import games.strategy.triplea.ai.proAI.data.ProPurchaseOption;
+import games.strategy.triplea.ai.proAI.data.ProTerritory;
 import games.strategy.triplea.attatchments.UnitAttachment;
 import games.strategy.triplea.attatchments.UnitSupportAttachment;
 import games.strategy.triplea.delegate.AirMovementValidator;
